BUILD: http-htx: fix build warning regarding long type in printf
Commit a66adf41e ("MINOR: http-htx: Add understandable errors for the
errorfiles parsing") added a warning when loading malformed error files,
but this warning may trigger another build warning due to the %lu format
used. Let's simply cast it for output since it's just used for end user
output.

This must be backported to 2.0 like the commit above.

+++ src/http_htx.c
@@ -970,7 +970,7 @@ int http_str_to_htx(struct buffer *buf, struct ist raw, char **errmsg)
 	}
 	if ((flags & HTX_SL_F_CLEN) && h1m.body_len != (raw.len - ret)) {
 		memprintf(errmsg, "payload size does not match the announced content-length (%lu != %lu)",
-			  (raw.len - ret), h1m.body_len);
+			  (unsigned long)(raw.len - ret), (unsigned long)h1m.body_len);
 		goto error;
 	}
